Server/requirements.txt
flake8==4.0.1
pytest-cov==3.0.0
opencv-python==4.6.0.66
tensorflow==2.9.1
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
distlib==0.3.4
filelock==3.6.0
firebase-admin==5.2.0
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.33.3
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
imageio==2.19.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
msgpack==1.0.4
networkx==2.8.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
proto-plus==1.20.6

protobuf==3.20.1

psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.1
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.0.3
tifffile==2022.5.4
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.10
virtualenv==20.14.1
Werkzeug==2.1.2
